Relax validation for AttrDtype
Before the refactoring, we already allowed using Python build-in types
to be used in relation to attributes (e.g. op.cast(inp, to=str)). We
only explicitly forbid using `object`. However, `object` is a rather
common dtype to come across when using pandas with string
columns. Furthermore, we already use python build in types in the
quantcore.learn converters. It is natural to allow them since users
coming from numpy probably expect that behavior. This PR is needed to
make the existing converters work on origin/main again.