Build graphs with an explicit IR version
onnxruntime is not yet able to consume models with the latest IR
version. Until the ecosystem had a reasonable amount of time to adopt
the new IR version we should not produces models with it.

Setting the IR version like this is a safe way according to an
upstream discussion: https://github.com/onnx/onnx/discussions/5203#discussioncomment-5841152